# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BusinessOS API for Rose Beauty Palace
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product & inventory management
  - name: purchases
    description: Purchasing & supply chain
  - name: sales
    description: Point of sale
  - name: campaigns
    description: Marketing & ROI tracking
  - name: dashboard
    description: Dashboard metrics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get dashboard metrics
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Dashboard metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardMetrics"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: includeArchived
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/{sku}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by SKU
      parameters:
        - name: sku
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: sku
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductRequest"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: archiveProduct
      tags: [products]
      summary: Archive a product (soft delete)
      parameters:
        - name: sku
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product archived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /purchases:
    get:
      operationId: listPurchases
      tags: [purchases]
      summary: List all purchases
      responses:
        "200":
          description: List of purchases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Purchase"
    post:
      operationId: createPurchase
      tags: [purchases]
      summary: Log a new purchase batch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePurchaseRequest"
      responses:
        "201":
          description: Purchase created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"

  /purchases/{id}/arrive:
    post:
      operationId: markPurchaseArrived
      tags: [purchases]
      summary: Mark a purchase as arrived (adds to inventory)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase marked arrived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /sales:
    get:
      operationId: listSales
      tags: [sales]
      summary: List all sales
      parameters:
        - name: campaignId
          in: query
          schema:
            type: integer
        - name: days
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
    post:
      operationId: createSale
      tags: [sales]
      summary: Log a new sale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSaleRequest"
      responses:
        "201":
          description: Sale created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
        "400":
          description: Insufficient stock or invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List all campaigns
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a new campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignRequest"
      responses:
        "201":
          description: Campaign created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"

  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get campaign with ROI details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign with ROI
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignWithROI"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message

    DashboardMetrics:
      type: object
      properties:
        netProfit:
          type: integer
          description: Net profit in cents (TZS * 100)
        totalRevenue:
          type: integer
          description: Total revenue in cents
        totalCogs:
          type: integer
          description: Total cost of goods sold in cents
        totalMarketingSpend:
          type: integer
          description: Total marketing spend in cents
        totalInventoryValue:
          type: integer
          description: Total inventory value in cents
        lowStockItems:
          type: array
          items:
            $ref: "#/components/schemas/LowStockItem"
        topSellers:
          type: array
          items:
            $ref: "#/components/schemas/TopSeller"
        days:
          type: integer
      required:
        - netProfit
        - totalRevenue
        - totalCogs
        - totalMarketingSpend
        - totalInventoryValue
        - lowStockItems
        - topSellers
        - days

    LowStockItem:
      type: object
      properties:
        sku:
          type: string
        name:
          type: string
        currentStock:
          type: integer
        reorderThreshold:
          type: integer
      required:
        - sku
        - name
        - currentStock
        - reorderThreshold

    TopSeller:
      type: object
      properties:
        sku:
          type: string
        name:
          type: string
        unitsSold:
          type: integer
        revenue:
          type: integer
      required:
        - sku
        - name
        - unitsSold
        - revenue

    Product:
      type: object
      properties:
        sku:
          type: string
        name:
          type: string
        category:
          type: string
        baseUnitName:
          type: string
        sellingPrice:
          type: integer
          description: Selling price in cents (TZS * 100)
        currentStock:
          type: integer
        reorderThreshold:
          type: integer
        averageLandedCost:
          type: integer
          description: Weighted average landed cost in cents
        profitMarginPercent:
          type: number
          description: Profit margin percentage
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - sku
        - name
        - category
        - baseUnitName
        - sellingPrice
        - currentStock
        - reorderThreshold
        - averageLandedCost
        - profitMarginPercent
        - isActive
        - createdAt

    CreateProductRequest:
      type: object
      properties:
        sku:
          type: string
        name:
          type: string
        category:
          type: string
        baseUnitName:
          type: string
        sellingPrice:
          type: integer
          description: Selling price in cents (TZS * 100)
        reorderThreshold:
          type: integer
      required:
        - sku
        - name
        - category
        - baseUnitName
        - sellingPrice
        - reorderThreshold

    UpdateProductRequest:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        baseUnitName:
          type: string
        sellingPrice:
          type: integer
        reorderThreshold:
          type: integer

    Purchase:
      type: object
      properties:
        id:
          type: integer
        sku:
          type: string
        productName:
          type: string
        supplier:
          type: string
        quantityBaseUnits:
          type: integer
        totalProductCost:
          type: integer
          description: In cents
        transportCost:
          type: integer
          description: In cents
        landedUnitCost:
          type: integer
          description: Calculated landed cost per unit in cents
        status:
          type: string
          enum: [pending, arrived]
        purchaseDate:
          type: string
          format: date-time
        expectedArrivalDate:
          type: string
          nullable: true
        arrivedAt:
          type: string
          nullable: true
          format: date-time
      required:
        - id
        - sku
        - productName
        - supplier
        - quantityBaseUnits
        - totalProductCost
        - transportCost
        - landedUnitCost
        - status
        - purchaseDate

    CreatePurchaseRequest:
      type: object
      properties:
        sku:
          type: string
        supplier:
          type: string
        quantityBaseUnits:
          type: integer
        totalProductCost:
          type: integer
          description: In cents (TZS * 100)
        transportCost:
          type: integer
          description: In cents (TZS * 100)
        purchaseDate:
          type: string
          format: date-time
        expectedArrivalDate:
          type: string
          nullable: true
      required:
        - sku
        - supplier
        - quantityBaseUnits
        - totalProductCost
        - transportCost
        - purchaseDate

    Sale:
      type: object
      properties:
        id:
          type: integer
        sku:
          type: string
        productName:
          type: string
        quantitySold:
          type: integer
        totalCashReceived:
          type: integer
          description: In cents
        costOfGoods:
          type: integer
          description: In cents
        profit:
          type: integer
          description: In cents
        campaignId:
          type: integer
          nullable: true
        campaignName:
          type: string
          nullable: true
        saleDate:
          type: string
          format: date-time
      required:
        - id
        - sku
        - productName
        - quantitySold
        - totalCashReceived
        - costOfGoods
        - profit
        - saleDate

    CreateSaleRequest:
      type: object
      properties:
        sku:
          type: string
        quantitySold:
          type: integer
        totalCashReceived:
          type: integer
          description: In cents (TZS * 100)
        campaignId:
          type: integer
          nullable: true
        saleDate:
          type: string
          format: date-time
      required:
        - sku
        - quantitySold
        - totalCashReceived
        - saleDate

    Campaign:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        adSpend:
          type: integer
          description: In cents
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          nullable: true
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - adSpend
        - startDate
        - createdAt

    CreateCampaignRequest:
      type: object
      properties:
        name:
          type: string
        adSpend:
          type: integer
          description: In cents (TZS * 100)
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          nullable: true
          format: date-time
      required:
        - name
        - adSpend
        - startDate

    CampaignWithROI:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        adSpend:
          type: integer
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          nullable: true
          format: date-time
        totalRevenue:
          type: integer
        totalProfit:
          type: integer
        unitsSold:
          type: integer
        costPerAcquisition:
          type: integer
          description: CPA in cents
        roi:
          type: number
          description: ROI percentage
        sales:
          type: array
          items:
            $ref: "#/components/schemas/Sale"
      required:
        - id
        - name
        - adSpend
        - startDate
        - totalRevenue
        - totalProfit
        - unitsSold
        - costPerAcquisition
        - roi
        - sales