# lib/api-zod/src/generated/types/campaignWithROI.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BusinessOS API for Rose Beauty Palace
 * OpenAPI spec version: 0.1.0
 */
import type { Sale } from "./sale";

export interface CampaignWithROI {
  id: number;
  name: string;
  adSpend: number;
  startDate: Date;
  endDate?: Date | null;
  totalRevenue: number;
  totalProfit: number;
  unitsSold: number;
  /** CPA in cents */
  costPerAcquisition: number;
  /** ROI percentage */
  roi: number;
  sales: Sale[];
}