# lib/api-zod/src/generated/types/createCampaignRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BusinessOS API for Rose Beauty Palace
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCampaignRequest {
  name: string;
  /** In cents (TZS * 100) */
  adSpend: number;
  startDate: Date;
  endDate?: Date | null;
}