# lib/api-zod/src/generated/types/createProductRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BusinessOS API for Rose Beauty Palace
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProductRequest {
  sku: string;
  name: string;
  category: string;
  baseUnitName: string;
  /** Selling price in cents (TZS * 100) */
  sellingPrice: number;
  reorderThreshold: number;
}