# lib/api-zod/src/generated/types/createPurchaseRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BusinessOS API for Rose Beauty Palace
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePurchaseRequest {
  sku: string;
  supplier: string;
  quantityBaseUnits: number;
  /** In cents (TZS * 100) */
  totalProductCost: number;
  /** In cents (TZS * 100) */
  transportCost: number;
  purchaseDate: Date;
  expectedArrivalDate?: string | null;
}